Add pix reading
A doozy.

* Move `framebuffer` into the VM object (it wasn't there before because it
didn't need to be there before)
* Add `screen` and `screen_dirty` to the VM object (`Image` representing screen
content and flag for when image is out of date resp.)
* When reading pixels from the screen, fetch/refetch framebuffer contents to
the screen `Image` so we can `GetImageColor(vm.screen, x, y)`.
* When setting pixels onto the screen, if we have a valid `screen` object then
keep it updated.
* For pretty much every drawing function (just `cls` and `print` at the moment)
determine the tradeoff between keeping the screen up-to-date or simply
setting the dirty flag and letting the read performance suffer. (So far, `cls`
is probably fine to `ImageClearBackground` at 320x240, while `print` is
probably more complicated and it'd be easier for my sake to just not; if I
decide to change this around at some later date I won't be too surprised.)